Movido todo lo referente a la renovación del jwt a la estructura auth

+++ src/services/user.service.ts
@@ -1,5 +1,7 @@
 import { JwtPayload } from "jsonwebtoken";
 import { ValidationError } from "sequelize";
+import { sequelize } from "../database/database.conexion";
+import CustomErrorClass from "../errors/custom-error.class";
 
 import { compare, encrypt } from "../helpers/bcrypt.helper";
 import { deleteFile } from "../helpers/delete-img.helper";
@@ -11,74 +13,7 @@ import { passwordModel } from "../models/password.model";
 import { userModel } from "../models/user.model";
 
 class UserService {
-    //ToDo: mover los métodos login, register, renewToken y deleteAccount a authController.
-    /**
-     * Checks if the email and password passed by parameter are found in the database.
-     * 
-     * If the email is found, it compares the encrypted password with the one passed by parameter.
-     * 
-     * If both data are correct, it generates a JWT and returns a response of type UserResponse.
-     * 
-     * @param email ```string```
-     * @param password ```string```
-     * @returns ```Promise<UserResponse>```
-     */
-    async login(email: string, password: string): Promise<UserResponse> {
-        let cryptPass: boolean = false;
-        try {
-            const user: any = await userModel.findOne({ where: { email } });
 
-            // Compare passwords.
-            if (user !== null) cryptPass = compare(password, user?.password);
-
-            // If any data is invalid, it throws an error.
-            if (!cryptPass || user === null) {
-                const message = 'Usuario o contraseña incorrectos.';
-                const err = new Error(message) as CustomError;
-                err.status = 404;
-                throw err;
-            }
-
-            // Generate JWT.
-            const token = generateJwt(user.id, user.username, user.img);
-
-            return { status: 'OK', user, token };
-
-        } catch (error) {
-            throw error;
-        }
-    }
-
-    /**
-     * Registers a new user in the system.
-     * 
-     * @param name - The name of the user.
-     * @param username - The username of the user.
-     * @param email - The email address of the user.
-     * @param password - The password of the user.
-     * @param img - The URL of the user's profile image.
-     * @returns An object containing the new user, a JWT token and a status message.
-     * @throws {CustomError} When there's a validation error (status 422) or an unexpected error (status 500).
-     */
-    async register(name: string, username: string, email: string, password: string, img: string): Promise<any> {
-        try {
-            const passBcrypt = encrypt(password); //Encrypt the password
-            const newUser: any = await userModel.create({ name, username, email, password: passBcrypt, img });
-
-            const token = generateJwt(newUser.id, newUser.username, newUser.img); //Generate JWT
-            return { status: 'OK', user: newUser, token }
-
-        } catch (error) {
-            if (error instanceof ValidationError) {
-                const message = error.errors[0]?.message || 'Error en la validación del formulario';
-                const err = new Error(message) as CustomError;
-                err.status = 422;
-                throw err;
-            }
-
-            throw new Error('Unexpected error');
-        }
-    }
 
     async updateUser(id: string, name: string, username: string, email: string) {
         try {
@@ -144,51 +79,45 @@ class UserService {
         }
     }
 
-    async renewToken(token: string) {
-        if (!token) {
-            const message = 'Error en el token de acceso';
-            const err = new Error(message) as CustomError;
-            err.status = 401;
-            throw err;
-        }
 
-        try {
-            const { id, username, img } = validateJWT(token) as JwtPayload;
-            const newToken = generateJwt(id, username, img);
-            const user = await userModel.findByPk(id);
-            return { status: 'OK', user, token: newToken };
-
-        } catch (error) {
-            if (error instanceof ValidationError) {
-                const message = error.errors[0]?.message || 'Err al actualizar la contraseña';
-                const err = new Error(message) as CustomError;
-                err.status = 404;
-                throw err;
-            }
-            throw new Error('Unexpected error');
-        }
-    }
-
+    /**
+    * Deletes an user account and all the related data from the database.
+    * 
+    * @param id The id of the user to be deleted.
+    * @throws {CustomErrorClass} If the user data is invalid or not authorized.
+    * @throws {Error} If an unexpected error occurs.
+    */
     async deleteAccount(id: string) {
+        // Start a managed transaction.
+        const dbTransaction = await sequelize.transaction();
+
         try {
+            // Find the user by id.
             const user: any = await userModel.findByPk(id);
 
-            if (user.img !== 'uploads/user.png') {
-                deleteFile(user.img)
-            };
+            // Delete the user image file if it is not the default one.
+            if (user.img !== 'uploads/user.png') deleteFile(user.img);
 
-            await passwordModel.destroy({ where: { userId: id } });
-            await categoryModel.destroy({ where: { userId: id } });
-            await userModel.destroy({ where: { id } });
+            // Create an array of promises to delete the passwords, categories and user records.
+            const promises = [
+                passwordModel.destroy({ where: { userId: id }, transaction: dbTransaction }),
+                categoryModel.destroy({ where: { userId: id }, transaction: dbTransaction }),
+                userModel.destroy({ where: { id }, transaction: dbTransaction }),
+            ];
+
+            // Execute the promises in parallel.
+            await Promise.all(promises);
+
+            // Commit the transaction.
+            await dbTransaction.commit();
 
         } catch (error) {
-            if (error instanceof ValidationError) {
-                const message = error.errors[0]?.message || 'Err al actualizar la contraseña';
-                const err = new Error(message) as CustomError;
-                err.status = 404;
-                throw err;
-            }
-            throw new Error('Unexpected error');
+            // Rollback the transaction.
+            await dbTransaction.rollback();
+
+            // Errors.
+            if (error instanceof ValidationError) throw new CustomErrorClass(error.errors[0]?.message, 400);
+            throw new Error('Unexpected error.');
         }
     }
 }
